feat: all links created by user are deleted if that user's account is deleted

+++ server/controllers/users/deleteUser.js
@@ -1,4 +1,5 @@
 const { User } = require("../../models/User");
+const { Link } = require("../../models/Link");
 const { statusCodes } = require("../../utils/statusCodes");
 
 const deleteUserById = async (req, res) => {
@@ -17,6 +18,17 @@ const deleteUserById = async (req, res) => {
         .status(statusCodes.FORBIDDEN)
         .json({ message: "Invalid or incorrect user ID" });
       return;
+    }
+    const links = await Link.find({ userID });
+    if (!links) {
+      res
+        .status(statusCodes.FORBIDDEN)
+        .json({ message: "invalid or incorrect user ID" });
+      return;
+    }
+
+    for (let i = 0; i < links.length; i++) {
+      await Link.findOneAndRemove({ userID });
     }
     await User.findOneAndRemove({ userID });
     res
